Add display-metrics-changed handler to place window on screen on display matrics change
